php 32 bit compatibility

+++ src/BackupDestination/BackupDestination.php
@@ -139,7 +139,7 @@ class BackupDestination
         }
     }
 
-    public function usedStorage(): int
+    public function usedStorage(): float
     {
         return $this->backups()->size();
     }
